+++ packages/vue/src/svg/qr.vue
@@ -98,21 +98,6 @@ function generateQr() {
   if (props.moduleStyle && typeof props.moduleStyle === "function") {
     moduleStyleFunction = props.moduleStyle;
   }
-  // placing each modules in x,y position in the svg using fillRect
-  let path = { finder: "", codeword: "" };
-  let x = size;
-  let y = size;
-  for (let i = 0; i < qr.data.length; i++) {
-    const bit = qr.data[i];
-    if (bit) {
-      moduleStyleFunction(path, { x, y, size, index: i }, qr);
-    }
-    x += size;
-    if (i % qr.gridSize === qr.gridSize - 1) {
-      x = size;
-      y += size;
-    }
-  }
   // if image place the image in center, QR ErrorCorrectionLevel Should be high and Image should not be more that 25-30% of the qr size to scan the QR code properly
   if (props.image) {
     drawImageInCenter(qr, size, moduleStyleFunction);
@@ -169,7 +154,7 @@ function drawImageInCenter(
     canvas.height = height;
     const ctx = canvas.getContext("2d")!;
     ctx.globalAlpha = props.image!.opacity || DEFAULT_IMG_OPACITY;
-    ctx.drawImage(img, 0, 0, height, width);
+    ctx.drawImage(img, 0, 0, width, height);
     const a = props.image?.opacity || DEFAULT_IMG_OPACITY;
     imgData.value = { img: canvas.toDataURL(), height, width, x, y, a };
     // qr modules
@@ -238,9 +223,21 @@ function overlappingImage(
 }
 
 onMounted(() => generateQr());
-watch(props, () => {
-  generateQr();
-});
+watch(
+  [
+    () => props.input,
+    () => props.qrOptions?.errorCorrection,
+    () => props.size,
+    () => props.image?.src,
+    () => props.image?.opacity,
+    () => props.image?.sizePercent,
+    () => props.image?.overlap,
+    () => props.image?.border,
+  ],
+  () => {
+    generateQr();
+  }
+);
 defineExpose({ svgRef });
 </script>